Update FBScrapper.py
+ can skip task if file is exist
+ add profile name at the start of the file

+++ FBScrapper.py
@@ -63,6 +63,13 @@ def multipic_img_ids():
                                     5:current_url_end]
     return current_url_fix
 
+
+def click_next(current_url2):
+    img_id_list.append(current_url2)
+    # print(current_url2)
+    driver.find_element_by_xpath(
+        '/html/body/div/div/div[2]/div/div[1]/div/div/div[1]/div/div[2]/table/tbody/tr/td[2]/a').click()
+
 # press the okay button if it get blocked by facebook
 time.sleep(2)
 if check_element_by_link_text('Okay'):
@@ -75,11 +82,33 @@ last_height = driver.execute_script("return document.body.scrollHeight")
 poster_count = 1
 multipic_count = 5
 timestamp_count = 0
+title = str(driver.title).replace(" ", "_")
 while True:
-    timestamp = get_timestamp(timestamp_count)
+    timestamp_elements = driver.find_elements_by_tag_name('abbr')
 
+    # post with single image
+    posts = driver.find_elements_by_class_name(
+        "_39pi")
     # post with multiple image
     multipic_url = driver.find_elements_by_class_name("_26ih")
+
+    if (poster_count-1 == len(posts)-1) and (len(multipic_url)-1 == multipic_count-1):
+        # Scroll down to bottom
+        driver.execute_script(
+            "window.scrollTo(0, document.body.scrollHeight);")
+        # Wait to load page
+        time.sleep(3)
+        # Calculate new scroll height and compare with last scroll height
+        new_height = driver.execute_script("return document.body.scrollHeight")
+        if new_height == last_height:
+            print('we\'re done here this is the last pages!')
+            break
+        last_height = new_height
+        continue
+
+    timestamp = get_timestamp(timestamp_count)
+    file_path_single = Path(f'./image/{title}_{timestamp}.jpg')
+
     # print('nope it still', len(multipic_url))
     if len(multipic_url) > multipic_count:
         multipic_count = len(multipic_url)
@@ -102,24 +131,27 @@ while True:
         img_id_list = list()
         img_count = 0
         while True:
+            file_path_multipic = Path(
+                f'./image/{title}_{timestamp}_p{img_count}.jpg')
             current_url2 = multipic_img_ids()
             if current_url2 in img_id_list:
                 print('last image, done!')
                 break
-            filename_format = f'./image/{timestamp}_p{img_count}.jpg'
-            img_id_list.append(current_url2)
-            # print(current_url2)
-            driver.find_element_by_xpath(
-                '/html/body/div/div/div[2]/div/div[1]/div/div/div[1]/div/div[2]/table/tbody/tr/td[2]/a').click()
+            if file_path_multipic.is_file():
+                print(f'skipping {file_path_multipic} file exist')
+                click_next(current_url2)
+                img_count = img_count + 1
+                continue
+            click_next(current_url2)
             fullsize_img_url = driver.find_element_by_xpath(
                 '/html/body/div/div/div[2]/div/div[1]/div/div/div[3]/div[1]/div[2]/span/div/span/a[1]').get_attribute('href')
-            print('saving multi', filename_format)
+            print('saving multi', file_path_multipic)
             driver.get(fullsize_img_url)
             image_url = driver.current_url
             # print(image_url) # image url
             driver.back()
             urllib.request.urlretrieve(
-                image_url, filename_format)  # Download Image
+                image_url, file_path_multipic)  # Download Image
             img_count = img_count + 1
             time.sleep(0.5)
         close_new_tab()
@@ -127,35 +159,24 @@ while True:
         driver.back()
         timestamp_count = timestamp_count + 1
     else:  # post with single image
-        posts = driver.find_elements_by_class_name(
-            "_39pi")
+        if file_path_single.is_file():
+            print(f'skipping {file_path_single} file exist', timestamp_count)
+            poster_count = poster_count + 1
+            timestamp_count = timestamp_count + 1
+            continue
         poster = posts[poster_count]  # find the post with single image
         url = poster.get_attribute('href')  # get the post url
         # print(timestamp, '\n', url, '\n\n')
         open_new_tab(url)
-        timestamp_format = f'./image/{timestamp}.jpg'
         view_full_size = driver.find_element_by_link_text(
             "View Full Size").get_attribute('href')  # Get the image URL
         driver.get(view_full_size)
         image_url = driver.find_element_by_tag_name('img').get_attribute('src')
         # print(image_url)
         urllib.request.urlretrieve(
-        image_url, timestamp_format)  # Download Image
+            image_url, file_path_single)  # Download Image
         poster_count = poster_count + 1  # add 1 to the poster_count counter
-        print('saving single', timestamp_format)
+        print('saving single', file_path_single)
         time.sleep(0.5)
         close_new_tab()
-        time.sleep(0.5)
         timestamp_count = timestamp_count + 1
-
-    if (poster_count == len(posts)) and (len(multipic_url) == multipic_count):
-        # Scroll down to bottom
-        driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
-        # Wait to load page
-        time.sleep(3)
-        # Calculate new scroll height and compare with last scroll height
-        new_height = driver.execute_script("return document.body.scrollHeight")
-        if new_height == last_height:
-            print('we\'re done here this is the last pages!')
-            break
-        last_height = new_height
